fix: rename i18n module to jsx and update hooks
